sentinel: updateClusterView: update follow id only if master converged.
Now updateClusterView changes members follow id only when the master is
converged to the current clusterview.

This avoids members dropping their database and trying to sync (pg_basebackup)
to new master if this is not ready since it can bring to a state where no
members have data if the new elected master cannot promote itself or dies
before promoting.

Additionally it always returns a clusterView also if nothing has changed from
the previous one.