proxy: update to new pollon and use predefined check interval.
The proxy doesn't get the sleep interval, but now, it uses a low default
of 5 seconds for every check.

This tries to limit but doesn't avoid the risk of having a proxy with old clusterview and sending
connections to an unelected master.

In a possible future change the leader sentinel will wait for all the alive
proxies to be in sync before electing a new master.