keeper: check if master timeline is forked before our position.
If a standby is elected as the new master, the other standbys will switch syncing
to the new master. If for any reasons the elected standby was behind other standbys
it will fork a new timeline before the other standbys. The other standbys cannot
sync with it but will need a full resync (pg_basebackup at the moment).