proxy: stop listening on error or empty proxyConf.
Since multiple proxies can be load balanced, the lb needs to know which proxy
is healthy.

Before this patch on error communicating with etcd or with an empty proxyConf
the proxy closes connections to the previous master but continues accepting
connections and closing them instantly.

With this patch, on error communicating with etcd or with empty proxyConf the
proxy stops listening. It will restart listening on successful etcd
communication and a populated proxyConf.

In this way a load balancer with a standard tcp check will detect the proxy as
not healthy and not balance to it.

This is enabled by default and can be disabled with `--stop-listening=false`