Add InitWithMultipleKeepers cluster config option
When initing a new cluster (with an empty db) it can be useful to be able to
start multiple keepers and let the sentinel randomly choose one as the master
(since they are both empty).

This will avoid the need to start one keeper, start the sentinel(s), wait for
the sentinel to elect the keeper as master, then start the other keeper(s).

To make this useful on startup, there's the need to be able to specify an
initial cluster config, this will come in a future patch.