sentinel/keeper: handle store communication problems on start.
keeper: Since the clusterconfig is needed to instantiate the postgres manager,
loop trying to get it. This is needed since the postgres manager is currently
stateful and needs to be instantiated during the keeper startup (see the TODO
to make it stateless).

sentinel: the clusterconfig is needed at startup only to retrieve the
SleepInterval, if it cannot be retrieved just use the DefaultSleepInterval.