Fix problematic lock renewal after leader failure in Consul
This PR changes slightly the lock management during
the leader election process by removing the session
in a best effort fashion after a leader or store
failure.

This is necessary for Consul which restores the session
after the leader failure and leaves the current leader
unaware of its status change, thus leaving the cluster
without leader.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>