keeper: refactor user and password management and options
This patch tries to unify the user and password management.

* Since the replication username and password are also sensitive, remove
them from the clusterconfig and require the user to provide them as
keeper options. Obviously they need to have the same values for all the
keepers (`--pg-repl-username`, `--pg-repl-password[file]`).

* Use `--pg-su-username` and `pg-su-password[file]` options for all the
superuser required accesses:
 * to setup the superuser at db initialization time (previously it was
directly used the os username running the keepers, now it defaults to
this if not specified).
 * as credentials used by the keeper for superuser access to the postgres
instance.
 * as credentials for pg_rewind.

* The replication and superuser password options can be provided as
parameters (`--pg-su-password`, `--pg-repl-password`) or as a password
file (`--pg-su-passwordfile`, `--pg-repl-passwordfile`).