*: Only use PGParameters from the cluster spec.
With this patch the postgresql.conf will only include the defined
pgParameters in the cluster spec and will not include other config files
anymore.

This makes the postgres config handling more consistent and let us really
handle it in a centralized fashion without surprises.

When initializing a cluster we'd like to retain the previous pg
parameters:

* When initMode is new: the one created by initdb when initializing a new db
* When initMode is existing: the current instance parameters
* When handling point in time recovery: the parameters provided by the
restored instance.

For doing this, the keeper will, during initialization, rename and
include the existing `postgresql.conf`, get the list of the modified
parameters, report them to the sentinel that will merge them back in the
cluster spec.