*: Implement point in time recovery init mode.
Add a `pitr` initMode and related pitrConfig to the cluster spec.
When a cluster is initialized in this mode it'll do a point in time
recovery.

At first it will execute the `dataRestoreCommand` to restore the db
cluster data dir, then, if successfull, it'll setup a `recovery.conf`
file with the provided archive `restoreCommand` and start the instance,
if the start command exit with success then the recovery has been
completed.

In future additional option to define the various recovery target settings
(https://www.postgresql.org/docs/9.6/static/recovery-target-settings.html)
could be added.