*: report keeper and db states via store
The keeper and db states are reported using the store.

The keeper doesn't listen on a tcp port anymore since there isn't the
need to expose an api to query its state. The related options have been
removed (--port and --listen-address).

In addition, the kubernetes discovery has been removed since everything
is done using the store. The kubernetes discovery was quite complicated
and probably useful only for having the pod list from k8s and skip old
partitioned pods able to talk with the store. Probably this case is very
special and doesn't requires this sort of handling and can be generally
improved in future with additional checks (as a last resort a check that
does a k8s based filtering could be implemented).
Also the related sentinel options have been removed.