sentinel: wait for proxies closing connections
This patch waits for proxies to report that they closed connections to
an unelected master before enabling connections to the new one.

This tries to avoid a small windows where a proxy can send data to
an unelected master when a new master is already active and other
proxies are already sending connection to it causing some transactions to
go to the right master while other to the old one since the proxy hasn't
yet read from the store to close its connections.

This could happen only when the failover is faster than the proxy
timeout and this isn't usually the case.

This is done reading the proxyInfo reported by the proxy to check if
they are at the right proxy generation.

If a proxy isn't reporting its proxyInfo (problems
talking to the store, partitioned etc...) the proxyInfo will expire and
the sentinel will go ahead assuming that the proxy has closed its
connection (due to internal proxy timeout).

This is the best we can do and it's done assuming that the proxy is
behaving correctly closing the connections after the proxy timeout. One
case that could break this rule is when the host is very loaded and the
processes are starved causing the timeout fire and pollon closing the
connection after more time, but in this case we can assume that also the
connection proxying should be starved in the same way. This can be
called as a self-fencing. The unique way to be sure that this won't
happen will be to do real fencing of the proxy (for example killing
the proxy process).

In addition only proxies marked as enabled in the cluster data will
proxy connections. This check enforces that only proxy already known by
the sentinel and marked by it as enabled will proxy connections, in this
way if during election of a new master a new proxy starts and reads from
the cluster data to proxy to the unelected master it will not do this
since the sentinel haven't marked it as enabled.