*: add support for etcd v3 api
for etcdv2 and consul we keep using libkv while for etcd v3 api we'll directly
use the etcdclientv3. We abstract the implementation under a KVStore and an
Election interface. We haven't implemented a libkv store for etcdv3 since libkv
looks quite dead and doesn't support the context package. So in future we'll
probably move out from libkv to better use context features.