sentinel: handle stale proxyInfos
Don't rely on proxyInfos being expired by the store but handle this at the
sentinel level. Like already done for the keepers check that the published
proxyInfo is updated by checking that the published InfoUID is changed or ignore
it if it's not updated for a specified time interval (assuming that the proxy
stopped listening in this interval).

This is also needed for future kubernetes based store where the proxyInfos is
read from the pod and under some error conditions it won't expire but just not
being updated.

Also improve the proxy spec update logic in updateCluster to only check for
converged proxies when a new master is elected (the proxies spec master is
empty)