sentinel: improve sync standby management logic
* Update the sync standbys only if the reported ones are the same as the required
ones. In this way, when we have to choose a new master we are sure that there're
no intermediate changes between the reported standbys and the required ones.

* When we have to choose a new sync standby as master we can just choose one of
the common standbys between the reported and the required one since we are sure
these are defined in the master as sync standbys

* Improve logic to always have some real sync standbys up to
MinSynchronousStanbys, also if not in good state, to improve the possibility
to be able to choose a sync standby to replace a failed master.

* Always keep a common standby between the previous and the wanted
synchronousStandbys