*: Add a command to force fail a keeper
Add a `failkeeper` command to `stolonctl`. This command wil force a keeper as
"temporarily" failed.

It's just a one shot operation, the sentinel compute a new clusterdata
considering the keeper as failed and then restore its state to the real one.

This can be useful to force a new master keeper election. For example, if the
force failed keeper is a master, the sentinel will try to elect a new master, if
no new master can be elected the force failed keeper, if really healthy, will be
re-elected as master`,