Make DBWaitReadyTimeout configurable
This value is currently hardcoded to 60 seconds, and it is the time that
stolon will wait for Postgres to boot. In some circumstances (e.g. if
there is a lot of WAL to recover) Postgres can take longer than 60
seconds to boot. Allowing this value to be configured lets users cater
for that scenario. We do this by adding it to the cluster spec, so that
this value is consistent across all nodes in the cluster.