Fix for hanging sentinels
If etcd returned an error, our previous behaviour would be to report the
error and restart an election. The outer loop never responded to the
error by stopping the election, which would cancel the context, and
instead would restart a new campaign.

The first campaigns session will continue to exist on the context the
election loop had originally provided it, which is context.Background.
We'll come round to perform a new leadership election and the
etcdElection.Campaign() method will block waiting for all existing
leased keys to be removed from the prefix. Because the old campaign
sessions are still alive, they'll continue to renew their leases and
ensure those revisions never get deleted, thus blocking our sentinel
campaign loop indefinitely.

After discussion in #659, it became clear that most of the store
implementations are likely to have similar issues, and instead of fixing
the underlying implementations we decided to amend the interface to warn
consumers that they need to terminate existing elections before starting
new ones.

This is a sharp edge of the election API and one that might catch us in
a future refactor. This is balanced against a hope that we'll refactor
the election interface before a second use case is likely to arise.