Enable progress reporting for pg_basebackup
This will cause pg_basebackup to output some extra steps during startup
and shutdown. If progress reporting is also enabled (-P) then it will
show the exact file name that is currently being processed. This can be
useful for debugging and monitoring the backup process.

Combined with verbose mode (-v) this will give detailed information on
what files are being processed and how far through the backup we are.
There's a slight initial time cost introduced as Postgres has to
determine the total database size, but this should be minimal compared
to the backup time.

Co-authored-by: me@lawrencejones.dev
Co-authored-by: harrymaclean@gmail.com