Provide stolonctl environment variables to shells
Fixes #13.  We want to set the environment variables
stolonctl expects (STOLONCTL_CLUSTER_NAME, etc.) on shells
so users can fly ssh console in and run stolonctl to more
easily administrate their cluster.

Those environment variables are written to a file at
/data/.env by the "start" Go program after the defaults
are set.

Unfortunately, the shell that is created by hallpass appears to be
hard coded to /bin/sh, is not a login shell nor controllable by
/etc/passwd, so our best option is to use the POSIX ENV variable.